third_party/WebKit/LayoutTests/fast/dom/custom/custom-elements-registry-expected.txt

This is a testharness.js-based test.
PASS window.customElements.define is defined
FAIL window.customElements.define requires two arguments Test bug: need to pass exception to assert_throws()
Harness: the test ran to completion.
